+++ src/common/models/states.types.helper.ts
@@ -1,0 +1,9 @@
+import { IPaginationState, IQueryPaginationState, IBodyPaginationState } from "./states.model";
+
+export const isQueryPaginationState = (obj: IPaginationState): obj is IQueryPaginationState => {
+    return obj.source === 'query';
+}
+
+export const isBodyPaginationState = (obj: IPaginationState): obj is IBodyPaginationState => {
+    return obj.source === 'body';
+}
